requirements.txt
absl-py==2.3.1
astunparse==1.6.3
beautifulsoup4==4.13.5
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
contourpy==1.3.3
cycler==0.12.1
ffmpeg-python==0.2.0
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.59.2
fsspec==2025.9.0
future==1.0.0
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.74.0
gTTS==2.5.4
h5py==3.14.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
keras==3.11.3
kiwisolver==1.4.9
libclang==18.1.1
llvmlite==0.44.0
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.3
more-itertools==10.8.0
mpmath==1.3.0
namex==0.1.0
networkx==3.5
noisereduce==3.0.3
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai-whisper==20250625
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==11.3.0
protobuf==6.32.1
pycparser==2.23
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
pyttsx3==2.99
regex==2025.9.1
requests==2.32.5
rich==14.1.0
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
soundfile==0.13.1
soupsieve==2.8
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
tiktoken==0.11.0
torch==2.8.0
tqdm==4.67.1
triton==3.4.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
